// yanay-tevet-web/src/generated-files/api/workout-plan/types.gen.ts
export type WorkoutRoutineSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Order
     */
    order: number;
    /**
     * Exercise Count
     */
    exercise_count: number;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};
// ...
export type ExerciseSetSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Order
     */
    order: number;
    /**
     * Reps
     */
    reps: number | null;
    /**
     * Duration Seconds
     */
    duration_seconds: number | null;
};

/**
 * MeasurementType
 */
export type MeasurementType = 'reps' | 'time';
// ...
export type WorkoutExerciseSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Routine Id
     */
    routine_id: number;
    /**
     * Name
     */
    name: string;
    measurement_type: MeasurementType;
    /**
     * Notes
     */
    notes: string;
    /**
     * Order
     */
    order: number;
    /**
     * Sets
     */
    sets: Array<ExerciseSetSchema>;
    /**
     * Created At
     */
    created_at: string;
    /**
     * Updated At
     */
    updated_at: string;
};